# public/docs/_scribe/frontend/openapi.yaml
openapi: 3.0.3
info:
  title: 'MW Realty — Frontend API'
  description: 'REST API for the MW Realty mobile app: property search, listings, agents, content, customer accounts, enquiries and the AI assistant.'
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Get a token from <code>POST /api/auth/login</code>, <code>/api/auth/verify-otp</code> or <code>/api/auth/google</code>, and send it as <code>Authorization: Bearer {token}</code>.'
security:
  -
    default: []
paths:
...
  /api/auth/login:
    post:
      summary: 'Sign in'
      operationId: signIn
      description: 'Email + password sign-in. Returns a Bearer token for the `Authorization` header.'
      parameters:
        -
          in: header
          name: X-Device-Id
          description: ''
          example: '{DEVICE_ID}'
          schema:
            type: string
      responses:
        200:
          description: 'Signed in'
          content:
            application/json:
              schema:
                type: object
                example:
                  token: 12|9xQ...
                  token_type: Bearer
                  user:
                    id: 7
                    name: 'Sara Ahmed'
                    email: buyer@example.com
                    phone: '+971 50 123 4567'
                    location: 'Dubai Marina'
                    avatar_url: null
                    email_verified: true
                properties:
                  token:
                    type: string
                    example: 12|9xQ...
                  token_type:
                    type: string
                    example: Bearer
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 7
                      name:
                        type: string
                        example: 'Sara Ahmed'
                      email:
                        type: string
                        example: buyer@example.com
                      phone:
                        type: string
                        example: '+971 50 123 4567'
                      location:
                        type: string
                        example: 'Dubai Marina'
                      avatar_url:
                        type: string
                        example: null
                        nullable: true
                      email_verified:
                        type: boolean
                        example: true
        422:
          description: 'Wrong email or password'
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Invalid credentials.'
                properties:
                  message:
                    type: string
                    example: 'Invalid credentials.'
      tags:
        - 'User Auth'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: ''
                  example: buyer@example.com
                password:
                  type: string
                  description: ''
                  example: secret123
                device_name:
                  type: string
                  description: "A label for this device's token (shown nowhere yet, useful for support)."
                  example: 'iPhone 15 Pro'
                  nullable: true
              required:
                - email
                - password
      security: []
...
  /api/customer/me:
    get:
      summary: Dashboard
      operationId: dashboard
      description: "Everything for the account screen in one call: profile, counts, recent activity, the\nwishlist (property cards), saved searches and sent enquiries. For long wishlists, use the\npaginated `GET /api/customer/wishlist` instead."
      parameters:
        -
          in: query
          name: lang
          description: 'Language code for translated titles.'
          example: en
          required: false
          schema:
            type: string
            description: 'Language code for translated titles.'
            example: en
        -
          in: header
          name: X-Device-Id
          description: ''
          example: '{DEVICE_ID}'
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  profile:
                    name: 'Sara Ahmed'
                    email: buyer@example.com
                    phone: '+971 50 123 4567'
                    location: 'Dubai Marina'
                    avatar_url: null
                  stats:
                    wishlist: 2
                    saved_searches: 1
                    enquiries: 1
                  activity:
                    -
                      icon: heart.svg
                      text: 'You saved "Marina Gate 2BR" to your wishlist.'
                      time: '2 hours ago'
                  wishlist:
                    -
                      id: 12
                      slug: marina-gate-2br
                      image: 'https://.../photo.jpg'
                      ...: 'same card shape as GET /api/properties'
                  saved_searches:
                    -
                      id: 3
                      title: '2BR in Marina'
                      meta: 'Dubai Marina · 2 Bed · For sale'
                      criteria:
                        location: 'Dubai Marina'
                        bedrooms: '2'
                        listing_type: sale
                  enquiries:
                    -
                      id: 91
                      property: 'Marina Gate 2BR'
                      sent_to: 'Ahmed Khan'
                      date: 'Oct 08, 2026'
                      status: New
                properties:
                  profile:
                    type: object
                    properties:
                      name:
                        type: string
                        example: 'Sara Ahmed'
                      email:
                        type: string
                        example: buyer@example.com
                      phone:
                        type: string
                        example: '+971 50 123 4567'
                      location:
                        type: string
                        example: 'Dubai Marina'
                      avatar_url:
                        type: string
                        example: null
                        nullable: true
                  stats:
                    type: object
                    properties:
                      wishlist:
                        type: integer
                        example: 2
                      saved_searches:
                        type: integer
                        example: 1
                      enquiries:
                        type: integer
                        example: 1
                  activity:
                    type: array
                    example:
                      -
                        icon: heart.svg
                        text: 'You saved "Marina Gate 2BR" to your wishlist.'
                        time: '2 hours ago'
                    items:
                      type: object
                      properties:
                        icon:
                          type: string
                          example: heart.svg
                        text:
                          type: string
                          example: 'You saved "Marina Gate 2BR" to your wishlist.'
                        time:
                          type: string
                          example: '2 hours ago'
                  wishlist:
                    type: array
                    example:
                      -
                        id: 12
                        slug: marina-gate-2br
                        image: 'https://.../photo.jpg'
                        ...: 'same card shape as GET /api/properties'
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 12
                        slug:
                          type: string
                          example: marina-gate-2br
                        image:
                          type: string
                          example: 'https://.../photo.jpg'
                        ...:
                          type: string
                          example: 'same card shape as GET /api/properties'
                  saved_searches:
                    type: array
                    example:
                      -
                        id: 3
                        title: '2BR in Marina'
                        meta: 'Dubai Marina · 2 Bed · For sale'
                        criteria:
                          location: 'Dubai Marina'
                          bedrooms: '2'
                          listing_type: sale
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 3
                        title:
                          type: string
                          example: '2BR in Marina'
                        meta:
                          type: string
                          example: 'Dubai Marina · 2 Bed · For sale'
                        criteria:
                          type: object
                          properties:
                            location:
                              type: string
                              example: 'Dubai Marina'
                            bedrooms:
                              type: string
                              example: '2'
                            listing_type:
                              type: string
                              example: sale
                  enquiries:
                    type: array
                    example:
                      -
                        id: 91
                        property: 'Marina Gate 2BR'
                        sent_to: 'Ahmed Khan'
                        date: 'Oct 08, 2026'
                        status: New
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 91
                        property:
                          type: string
                          example: 'Marina Gate 2BR'
                        sent_to:
                          type: string
                          example: 'Ahmed Khan'
                        date:
                          type: string
                          example: 'Oct 08, 2026'
                        status:
                          type: string
                          example: New
      tags:
        - 'User Account'
...